sys,compilerJs: create static fields for closure Func parameters
Enhance JS compiler to create static fields for unique instances
of closure parameter lists and FuncType specifications. Use
those fields when creating the closure Func and invoking it.

This is an optimization aimed at reducing the overhead of creating
a closure Func and invoking it. We create the parameter lists and
FuncType once, and re-use it every time the closure is invoked.