web: fix redirect bug
After following a redirect the web client needs to return immediately.
Otherwise, the client attempts to again create an InStream for content,
when there is no more content.

In the case where the original request was GZIP encoded, when the GZIP
Input Stream is constructed, it immediately attempts to do some reads
to verify the gzip MAGIC NUMBER. Since there is no more content, this
fails.