domkit: move domkit css compilation downstream
This is the first step to support better customization and variable
support for domkit css.  Instead of shipping with the compiled CSS,
we'll package up the source CSS (fanCss); and introduce a new class
CompileCss that can be used downstream to assemble the source CSS
into the merged output.

At this stage the output is still exactly the same as the current
output CSS.