domkit: revert 7864a29d25d4 to restore staticallly compiled CSS
Now that we have made it further through redesigning the domkit CSS its
more clear that the burden of compiling CSS downstream outweighs having
to override a few more classes. This coupled with the docDomkit::Theming
chapter which gives clear examples of how to theme, it should be  much
simpler to just pull in the pod.res file.