domkit: Revert FilePicker.onSelect files.clear behavoir
We added a fix in 53b2d4f92b4e to clear our the input.files value, so that
if the same file(s) was re-selected the event handler would re-fire.

But this invalidates accessing FilePicker.files outside the onSelect callback
which is both really confusing and not terribly useful. So revert this back
to the original behavoir.

If this issue comes up again in practice we can revisit another solution...