js: Fix Buf crypto regression bug
Back in changeset f1c37f71a0f7 we changed the behavior of MemBuf.capacity to
default to 1024 to be consistent with JVM implementation.

This had the inadvertent effect of passing a bunch of empty bytes to all our
crypto funcs which produced the wrong hashes.

Fix is to trim raw byte array using Buf.size