dom:js: Fix Buf regression bug
Back in changeset f1c37f71a0f7 we changed the behavior of MemBuf.capacity to
default to 1024 to be consistent with JVM implementation.

This had the inadvertent effect of causing initialization of the view
Uint8 array to index out of bounds.

Fix is to trim raw byte array using Buf.size