domkit: Add Popup.fitBounds to verify overflow for async loaded content
Prior to this changeset; Popup only calculated overflow on the initial
open.  For content that may be async loaded and that may increase the
size of the popup; there was no mechanism to check the bounds again.

This patch pulls out that logic into a dedicated `Popup.fitBounds` method
that can be called by downstream consumers.