compilerJs: fix mixin inheritance
The compiler has taken care of resolving all slots to their implementations -
either a mixin or concrete class override. Previously, we were not making
use of that information to determine when to delegate a slot's implementation
to a mixin. In rare cases, this would incorrectly cause a mixin's implementation
to be used when a concrete override existed in the class hierarcy.