domkit: first pass on a fix for Table scrollwheel support in Firefox
Firefox has some trouble with wheel events interweaved with our repaint
events during a scroll.  Previously (and in other browsers) we used CSS
transitions to offload this to the GPU to keep scrolling smooth.  But at
some point this behavoir broke in Firefox, so this patch takes a stab at
a low level brute force attempt to throttle the wheel events to balance
a responsive scroll without being too choppy.