concurrent: make Future an abstract class
Previously Future was a final native class.  But its a general purpose API that has utility outside of the actor framework.  So in order to make it suitable for other use cases, make it abstract with little implementation.  Take the existing actor based implementation and create a new internal class ActorFuture.

We also had prototyped use cases for using Future as a "completable" to block threads until another thread called complete.  Under the covers, the ActorFuture suits this use case.  But create a new top-level factory called Future.makeCompletable for public use.