domkit: Temp fixes for Table event handling and bubbling
For Tables that contain child nodes like <a> links, and remove/create
them during repaint, the selection behavior would cause these to be
removed during event bubble.  So when the event continued upstream,
the original node may have been removed, nullifying the handlers.

This is most obvious when trying todo things like cmd/ctrl click a URL
in a <a> link in a table.

Technically you should not remove nodes like this inside this table
impl, but as its easy todo, and this behavior would be very difficult
to diagnose, take the hit internally to make sure it does not happen.

Longer term we need to rethink how this might work so we can update
the table focus/selection state more responsively.