domkit: improve Table support for <a> cell elems (fix for 160f9f6)
We previously had a temp fix (160f9f6) for handling of <a> elements inside
table cells.  This was never a long term solution, and seems to have
recently gotten flaky.

So this patch formalizes <a> link support directly in Table.  When we trap
mouse down events on <a> targets, the href gets directly invoked, bypassing
the normal render code path.  There is also a few edge cases with focus
that get updated to properly support clicking a link on an unfocused table.