sys: Formalize Env.var setup; add TimeZone/Locale.cur override via Env.var
Previously we relied on a backdoor hook from WebUtil.jsMain to set Env.vars.
However this occurred after all pod js code had already been run, which did
not give any code the opportunity to pick up env var configuration. We will
continue to support this behavoir, but it should be considered deprecated.

Env.var now looks for a `fan$env` object during construction which is used
to set the Env.vars as a proper immutable map at boot time.  The `fan$env`
object must occur _before_ `sys.js` in order to take effect.

Update TimeZone.cur and Locale.cur to support looking up the JsVM default
from Env.vars.  For TimeZone also add a primary fallback of querying the
browser for the current TimeZone to match the behavoir of JVM.  If one
cannot be determined, fallback to UTC.  This is a change from the existing
behavior which was hard-coded to New_York.