graphics: simplify to use one paint color for both stroke/fill; make stroke only width/pattern
This design brings graphics closer to original gfx design which used Brush + Pen.   It is not
quite as flexible/aligned with Canvas and SVG which maintains two separate colors for stroke and fill.
We will tackle that optimization in the pathing API.  But it is aligned with Java2D which just maintains one
color for both stroke/fill.  I think this is a simpler API because it's always confusing to think of text being a fill
operation and having to remember whether you set fill or stroke color for the text rendering.  Plus it
is more future proof because you can stroke outlines with gradients/patterns.