graphics: new GraphicsEnv API and public Image API
Previously we had a simple nodoc Image API that could parse png, jpeg, and svg to get basic meta-data such
as the natural size, color model, etc.  For png we had a bunch of meta-data required to render to PDFs.

As we flush out the Graphics rendering APIs we need a more comprehensive Image API which can be used with
JS Canvas, Java2D, buffered images, etc.  So we are creating a new GraphicsEnv class that models the toolkit
with three eventual implementations:
  - ServerGraphicsEnv: very simple headless support for image size and layout (PDF rendering)
  - DomGraphicsEnv: for use in browser and with CanvasGraphics
  - Java2DGraphicsEnv: backed by the java.awt APIs

As part of this refactor we move all the png, jpeg, and svg decoders into graphicsJava.  This gets them out
of the graphics API and trims down the JavaScript code for that pod.

In this changeset, I've marked that FontMetrics.metrics and DeviceContext should no longer be used.
The original FontData backing provided a simple solution to use predefined metrics data with only a DPI.
But going forward we will want to drive metrics strictly from a Graphics instance which is required to
support multiple displays with different DPIs, printing, etc.