domkit: remove last column pref width adjustment for header popup
This code is forcing a horizontal scrollbar on every table that actually tries to compute its width.  Which in turns
causes the last column data to always be clipped.  In reality the header popup only effects the column header
and shouldn't be applied to the overall column's preferred width.  I did examine using a separate headerWidth
in the model, but without breaking backward compatibility its a little tricky and doesn't really seem to effect the
quality of layout.