sys: enhance to Pod.readFPod to staged pods prior to loading
This is primarily useful for "hot" reloading of a pod that does not contain any fcode. When a pod->reload() happens,
it will close the FStore associated with the pod, forcing it to be re-read. If we detect that there is a staged
pod in ../install/<name>.pod then we install that first, and then load it.