compiler: remove error "Conflicting covariant returns"
I can't remember why I added this check and it appears to be limiting for no reason
as long as the method override is covariant fit for base type.  If this check does let
something slip thru, the runtime should still raise exception for cast exception.

Test code that illustrates where this was a problem is Haxall code base in the xeto
pod where CType.asm returns XetoSpec and AObj.asm returns Obj.  The ASpec.asm
method should be inherit both of those methods.