compilerEs: fix to properly emit static fields from mixins
Fix so that concrete types that implement mixins with static fields
have the static field accessors "copied" into the generated code
for that class.

Did not require any changes to t.ds generation because that was already
being handled by virtue of using the "implements" feature of TS to
indicate a concrete class implements a mixin.