compiler: fix dup closure shadow variables (take 2)
This is an enhancement to the fix yesterday where we were we are creating duplicate
local shadow variables in the closure method from the outer method. The first fix only
handled one level, but this fix handles multiple closure scopes.  We move the common
logic into a new MethodDef.getOrAddShadowVar method